requirements.txt
python-pptx
comtypes
PyPDF2